✨ Feat: 드롭다운 외부 클릭 시 닫힘 로직을 Navbar에서 Dropdown 컴포넌트로 이동
