+++ src/components/Navbar.tsx
@@ -54,35 +54,6 @@ export default function Navbar() {
     }
   }, [pathname]);
 
-  // 바깥 클릭 시 드롭다운 닫기
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      const target = event.target as Node;
-
-      // 드롭다운 메뉴 영역 확인
-      const dropdownMenus = document.querySelectorAll('[data-dropdown-menu]');
-      const isClickInsideDropdown = Array.from(dropdownMenus).some((menu) => menu.contains(target));
-
-      // 트리거 버튼인지 확인 (드롭다운을 토글하는 버튼들)
-      const isTriggerButton = (target as Element).closest('[data-dropdown-trigger]') !== null;
-
-      // navbar 영역 확인
-      const navbar = navbarRef.current;
-      const isClickInsideNavbar = navbar?.contains(target);
-
-      // 드롭다운 메뉴 영역 밖을 클릭하고, 트리거 버튼이 아닌 경우에만 닫기
-      if (!isClickInsideDropdown && (!isClickInsideNavbar || !isTriggerButton)) {
-        setOpenMenu(null);
-        setIsProfileDropdownOpen(false);
-      }
-    };
-
-    document.addEventListener('click', handleClickOutside);
-    return () => {
-      document.removeEventListener('click', handleClickOutside);
-    };
-  }, []);
-
   type MenuKey = keyof SidebarMenus;
 
   const menuConfigs: {
